src/main/java/com/crm/pages/LoginPage.java
package com.crm.pages;

import org.openqa.selenium.WebElement;
import org.openqa.selenium.support.FindBy;
import org.openqa.selenium.support.PageFactory;

import com.crm.base.TestBase_Driver;

public class LoginPage extends TestBase_Driver{
	
	
	@FindBy(name="username")
	WebElement username;
	
	
	@FindBy(name="password")
	WebElement password;
	
	@FindBy(xpath="//input[@type='submit']")
	WebElement loginBtn;
	
	@FindBy(xpath="//button[contains(text(),'Sign Up')]")
	WebElement signUpBtn;
	
	@FindBy(xpath="//img[contains(@class,'img-responsive')]")
	WebElement crmLogo;
	
	public LoginPage()
	{
		PageFactory.initElements(driver, this);
		
		
	}
	public String validateLoginPage()
	{
		return driver.getTitle();
	}
	
	public boolean validatecrmImgLogo()
	{
		return crmLogo.isDisplayed();
		
		
	}
	public HomePage Login(String un,String pwd)
	{
		username.sendKeys(un);
		password.sendKeys(pwd);
		loginBtn.click();
		return new HomePage();
		
		
		
	}
	
	
	
	
	
	
	
	
}
